pyproject.toml
[project]
name = "shoes-store-backend"
version = "0.1.0"
description = ""
authors = [
    {name = "AzRafa",email = "glowhithy@outlook.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django (>=5.2.8,<6.0.0)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "drf-nested-routers (>=0.95.0,<0.96.0)",
    "django-cors-headers (>=4.9.0,<5.0.0)",
    "djangorestframework-simplejwt (>=5.5.1,<6.0.0)",
    "pytest (>=9.0.1,<10.0.0)",
    "drf-yasg (>=1.21.11,<2.0.0)",
]

[tool.poetry]
packages = [{include = "shoes_store_backend", from = "src"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pytest-django (>=4.11.1,<5.0.0)"
]
